Completing dotnet install and sources for linux
