+++ Unix/install.py
@@ -27,6 +27,7 @@ import urllib.request
 
 config = {}
 releases = {}
+sources = {}
 
 
 def exception_config_missing():
@@ -89,8 +90,28 @@ def get_install_directory_from_config():
     new = config["directory"]
     if not new.endswith('/'):
         new = new + "/"
-
     return new
+
+
+def get_dotnet_destination():
+    return get_install_directory_from_config() + config["version"] + "/dotnet"
+
+
+def get_dotnet_core_path():
+    try:
+        return which("dotnet")
+    except:
+        if not os.path.isdir(get_dotnet_destination()):
+            tmp_path = "/tmp/spectero-dotnet-install.tar.gz"
+            link = get_dotnet_runtime_link()
+
+            # download, create the directory, extract.
+            os.system("wget %s -O %s -q" % (link, tmp_path))
+            os.system("mkdir -p %s" % get_dotnet_destination())
+            os.system("tar -xf %s -C %s" % (tmp_path, get_dotnet_destination()))
+
+        # Get the downloaded dotnet executable path.
+        return get_dotnet_destination() + "/dotnet"
 
 
 def get_download_channel_information():
@@ -176,10 +197,6 @@ def fix_permissions():
     os.system("usermod -m -d %s spectero" % get_install_directory_from_config()[:-1])
 
 
-def get_dotnet_core_path():
-    return which("dotnet")
-
-
 def create_systemd_service():
     if config["service"] == "true":
         try:
@@ -191,7 +208,7 @@ def create_systemd_service():
                 filedata = file.read()
 
             # Replace template data here
-            filedata = filedata.replace("ExecStart=/usr/bin/dotnet", "ExecStart=" + which("dotnet"))
+            filedata = filedata.replace("ExecStart=/usr/bin/dotnet", "ExecStart=" + get_dotnet_core_path())
 
             # Open a writer to the template location.
             with open(systemd_script_template, 'w') as file:
@@ -269,7 +286,7 @@ def create_shell_script():
         with open(cli_script, 'r') as file:
             filedata = file.read()
         print("Replacing variables in console management interface template...")
-        filedata = filedata.replace("{dotnet path}", which("dotnet"))
+        filedata = filedata.replace("{dotnet path}", get_dotnet_core_path())
         filedata = filedata.replace("{spectero working directory}", get_install_directory_from_config())
         filedata = filedata.replace("{version}", config["version"])
 
@@ -286,9 +303,27 @@ def create_shell_script():
         sys.exit(12)
 
 
+def get_sources_information():
+    global sources
+    try:
+        request = urllib.request.Request('https://raw.githubusercontent.com/ProjectSpectero/daemon-installers/master/SOURCES.json')
+        result = urllib.request.urlopen(request)
+        sources = json.loads(result.read().decode('utf-8'))
+    except:
+        request = urllib.request.Request('https://raw.githubusercontent.com/ProjectSpectero/daemon-installers/development/SOURCES.json')
+        result = urllib.request.urlopen(request)
+        sources = json.loads(result.read().decode('utf-8'))
+
+
+def get_dotnet_runtime_link():
+    global sources
+    return sources["linux"]["dotnet"]["x64"]
+
+
 if __name__ == "__main__":
     read_config()
     get_download_channel_information()
+    get_sources_information()
     validate_user_requests_against_releases()
     download_and_extract()
     create_user()
